Fixed a memory leak in pagetable_freepage().
We were not freeing the pagetableentry of the page that was to be freed.
Also, the pagetable entry was not updated to NULL to mark the page as
not allocated.

This would result in a page that was supposed to be freed to still be
marked as allocated in the page table. Along with that, we leak memory
by not freeing the pagetable entry.