Fixed pagetable_copy().
Earlier, pagetable_copy() was simply copying the page table entries
rather than copying the contents of the new pages. Now it allocates new
pages and then copies the contents of the old pages into the new ones.

Also, it now asks for the address space the new pages should belong to.