Implemented the address space functions.
This is the first iteration of the address space abstraction. It has not
yet been bug tested because vm_fault() is not yet implemented, so no
userspace process can run currently. Improvements will be made to this
implementation when the entire VM subsystem is able to run and testing
can be done.