Fixed incorrect index calculation in pagetables.
Many pagetable functions calculated the index into the page table
incorrectly. They zeroed out the bits not needed, but did not perform
the right shift needed. For example, the first level array uses the top
10 bits of the virtual address as the index of the page into the table.
Earlier, we zeroed out the bottom 22 bits but did not shift right by 22
bits so it resulted in the index being way larger then it should have
been.