Removed checks for valid segment addresses from seg_create().
If any of those checks failed, seg_create() returned NULL, which
resulted in an out of memory error being reported when in fact that is
not the case.

Now seg_create() fails only when there is no memory left and the checks
for whether all addresses in the segment are valid must be performed by
the calling function.