Fixed multiple bugs in as_define_stack().
-The stack segment was being created from the top of the userspace,
rather than the base of the stack. Because the segment starts at a lower
address and ends at a higher address, this would result in the stack
starting at the userspace top and ending somewhere in kseg0. Now the
stack starts at the base and ends at the userspace top.

-I was passing the total size of the stack to seg_create() rather than
passing the number of pages. This resulted in the stack being way bigger
than it should be. To be precise, the stack is 2M so that is 512 pages.
Previously I was setting the stack to be 2M pages.

-Earlier, I forgot to allocate the pages corresponding to the stack in
the page table. This resulted in a segmentation fault every time the
stack was accessed because the pages corresponding to the stack were
marked unallocated in the page table.