cm_freeupage now no longer requires the process freeing a page to own it.
There are some cases where the parent process calls proc_destroy() so
cm_freeupage() will fail in that situation. An example would be when we
run a program using fork() and the parent calls wait(), if the parent
does not exit before the child, the parent calls proc_destroy() when the
child exits. proc_destroy() calls pagetable_destroy() which in turn
calls cm_freeupage().

This check earlier resulted in leaked memory because the pages would
then never be freed.