Added a check in pagetable_destroy() to make sure the page it is freeing was actually allocated.
Since the current VM implementation uses lazy allocation, some virtual
pages are never allocated on physical memory. In that case, the physical
address assigned to that page is 0.

Calling cm_freeupage() *should* not actually do anything in this case
because it is very likely the coremap does not manage that page (the 0th
page is usually used by kernel during early boot). But still, make sure
no unwanted (and possibly very obscure) bugs creep in.