Fixed incorrect copying of segment structs in as_copy().
- as_copy() did not check for NULL pointers in the segment array, and
thus was causing the kernel to crash when any one of the segments was
NULL (a fairly common case).

- It was allocating different segments for as_stack and as_heap, when
those actually should be the same pointers as the ones in the segment
array.

- tempseg was allocated and assigned in the new segment array, resulting
in a valid pointer in that position, when the old segment array actually
contains NULL in the aforementioned position.